requirements.txt
psycopg2-binary==2.9.9
yt-dlp==2023.11.16